feat: centralize agent tasks
